+++ subgraphs/ethereum/src/liquidity/LiquidityUniswapV3.ts
@@ -1,15 +1,16 @@
 import { Address, BigDecimal, BigInt, log } from "@graphprotocol/graph-ts";
 
 import { TokenRecord, TokenSupply } from "../../../shared/generated/schema";
-import { BLOCKCHAIN, ERC20_OHM_V2, ERC20_TOKENS, getContractName, getWalletAddressesForContract, liquidityPairHasToken } from "../utils/Constants";
+import { BLOCKCHAIN, ERC20_OHM_V2, ERC20_TOKENS, getContractName, liquidityPairHasToken } from "../utils/Constants";
 import { getERC20DecimalBalance, getUniswapV3Pair } from "../utils/ContractHelper";
 import { getUSDRate } from "../utils/Price";
 import { toDecimal } from "../../../shared/src/utils/Decimals";
 import { getERC20Decimals } from "../contracts/ERC20";
 import { UniswapV3PositionManager } from "../../generated/ProtocolMetrics/UniswapV3PositionManager";
-import { createOrUpdateTokenRecord } from "../../../shared/src/utils/TokenRecordHelper";
-import { TYPE_LIQUIDITY, createOrUpdateTokenSupply } from "../../../shared/src/utils/TokenSupplyHelper";
+import { createTokenRecord } from "../../../shared/src/utils/TokenRecordHelper";
+import { TYPE_LIQUIDITY, createTokenSupply } from "../../../shared/src/utils/TokenSupplyHelper";
 import { TokenCategoryPOL } from "../../../shared/src/contracts/TokenDefinition";
+import { getWalletAddressesForContract } from "../utils/ProtocolAddresses";
 
 export const UNISWAP_V3_POSITION_MANAGER = "0xC36442b4a4522E871399CD717aBDD847Ab11FE88";
 const Q96 = BigInt.fromI32(2).pow(96);
@@ -184,7 +185,7 @@ export function getUniswapV3POLRecords(
       log.debug("getUniswapV3POLRecords: multiplier: {}", [multiplier.toString()]);
 
       records.push(
-        createOrUpdateTokenRecord(
+        createTokenRecord(
           timestamp,
           getContractName(pairAddress),
           pairAddress,
@@ -325,7 +326,7 @@ export function getUniswapV3OhmSupply(
       const ohmBalance = balances[ohmIndex];
 
       records.push(
-        createOrUpdateTokenSupply(
+        createTokenSupply(
           timestamp,
           getContractName(tokenAddress),
           tokenAddress,