+++ subgraphs/fantom/src/treasury/Assets.ts
@@ -1,10 +1,9 @@
-import { BigInt, log } from "@graphprotocol/graph-ts";
+import { BigInt, ethereum, log } from "@graphprotocol/graph-ts";
 
 import {
   TokenCategoryStable,
   TokenCategoryVolatile,
 } from "../../../shared/src/contracts/TokenDefinition";
-import { NewRound } from "../../generated/TokenRecords-fantom/ChainlinkAggregator";
 import { getOwnedLiquidityBalances } from "./OwnedLiquidity";
 import { getTokenBalances } from "./TokenBalances";
 
@@ -16,7 +15,7 @@ export function generateTokenRecords(timestamp: BigInt, blockNumber: BigInt): vo
   getOwnedLiquidityBalances(timestamp, blockNumber);
 }
 
-export function handleAssets(event: NewRound): void {
-  log.debug("handleAssets: *** Indexing block {}", [event.block.number.toString()]);
-  generateTokenRecords(event.block.timestamp, event.block.number);
+export function handleAssets(block: ethereum.Block): void {
+  log.debug("handleAssets: *** Indexing block {}", [block.number.toString()]);
+  generateTokenRecords(block.timestamp, block.number);
 }